+++ src/constants.ts
@@ -1,0 +1,2 @@
+// Mongo
+export const MONGO_URI = `mongodb://${process.env.MONGO_INITDB_ROOT_USERNAME}:${process.env.MONGO_INITDB_ROOT_PASSWORD}@bot-db:27017/${process.env.DATABASE_NAME}`;